Better fix for troubleshooting on windows
Changed the Troubleshooting section of README.md to include a more sane fix for the windows problem. 
Nim (1.6.6) at least already ships with libwinpthread-1.dll in mingw. The windows problem arises because libwinpthread-1.dll is not on PATH. 
Using static libraries means that we no longer need libwinpthread-1.dll to be on PATH.